+++ ReleaseNotes.md
@@ -14,6 +14,7 @@
 - improve DEFINE ANALYZER, now it uses predefined methods for tokenizers and filters, and has documentation for all of them
 - SELECT statement can use TEMPFILES clause on all levels except EXPLAIN (which is the last one)
 - update DEFINE USER statement
+- add UPSERT statement
 
 **Version 0.5.3 (compatible with SurrealDB version 1.5.2):**
 - minor fixes for docs and examples